+++ src/schema/generateFroidSchema.ts
@@ -10,10 +10,13 @@ import {
   parse,
 } from 'graphql';
 import {
+  FED2_DEFAULT_VERSION,
   ID_FIELD_NAME,
   EXTENDS_DIRECTIVE,
   KEY_DIRECTIVE,
   TAG_DIRECTIVE,
+  FED1_VERSION,
+  FED2_VERSION_PREFIX,
 } from './constants';
 import {implementsNodeInterface, externalDirective} from './astDefinitions';
 import {isRootType} from './isRootType';
@@ -27,6 +30,19 @@ import {createLinkSchemaExtension} from './createLinkSchemaExtension';
 import {createFederationV1TagDirectiveDefinition} from './createFederationV1TagDirectiveDefinition';
 import {ObjectTypeNode, KeyMappingRecord, ValidKeyDirective} from './types';
 import {removeInterfaceObjects} from './removeInterfaceObjects';
+
+type KeySorter = (keys: string[], node: ObjectTypeNode) => string[];
+
+const isEntity = (nodes: ObjectTypeNode[]): boolean =>
+  nodes.some((node) =>
+    node?.directives?.some(
+      (directive) => directive.name.value === KEY_DIRECTIVE
+    )
+  );
+
+const defaultKeySorter: KeySorter = (keys: string[]): string[] => {
+  return keys.sort((a, b) => a.indexOf('{') - b.indexOf('{'));
+};
 
 /**
  * Returns all non-root types and extended types
@@ -72,10 +88,12 @@ function getObjectDefinitions(
  * Returns all non-extended types with explicit ownership to a single subgraph
  *
  * @param {ObjectTypeDefinitionNode} node - The node to process `@key` directives for
+ * @param keySorter
  * @returns {ValidKeyDirective} The matching directive to use for generating the Global Object Identifier
  */
 function selectValidKeyDirective(
-  node: ObjectTypeDefinitionNode
+  node: ObjectTypeDefinitionNode,
+  keySorter: KeySorter
 ): ValidKeyDirective | undefined {
   const keyDirectives = node.directives?.filter(
     (directive) => directive.name.value === KEY_DIRECTIVE
@@ -85,17 +103,18 @@ function selectValidKeyDirective(
     return;
   }
 
-  // get field names that are @key to the entity
-  const firstValidKeyDirectiveFields = keyDirectives
+  // Prep the key directives for selection
+  const keys = keyDirectives
     .map((directive) => directive.arguments)
     .flat()
     .map((arg) => arg?.value as StringValueNode)
     .filter(Boolean)
     .map((strNode) => strNode.value)
     // Protect against people using the `id` field as an entity key
-    .filter((keys) => keys !== ID_FIELD_NAME)
-    .sort((a, b) => a.indexOf('{') - b.indexOf('{'))
-    .find((f) => f);
+    .filter((keys) => keys !== ID_FIELD_NAME);
+
+  // get field names that are @key to the entity
+  const firstValidKeyDirectiveFields = keySorter(keys, node).find((f) => f);
 
   if (!firstValidKeyDirectiveFields) {
     return;
@@ -173,33 +192,34 @@ function getTagDirectivesForIdField(
  * Generates key field nodes
  * Includes `@external` directive for Federation V1 generation
  *
- * @param {ObjectTypeDefinitionNode} node - The node to decorate
+ * @param {ObjectTypeDefinitionNode[]} nodes - The node to decorate
  * @param {KeyMappingRecord} keyMappingRecord - The list of key fields for the node
  * @param {FederationVersion} federationVersion - The version of federation to generate schema for
  * @returns {FieldDefinitionNode[]} A list field definitions
  */
 function getKeyFields(
-  node: ObjectTypeNode,
+  nodes: ObjectTypeNode[],
   keyMappingRecord: KeyMappingRecord,
   federationVersion: FederationVersion
 ): FieldDefinitionNode[] {
+  const nodeIsEntity = isEntity(nodes);
   const keyFieldNames = Object.keys(keyMappingRecord);
-
-  return (
-    node.fields
-      // take only @key fields and add @external directive to each of these
-      ?.filter((field) => keyFieldNames.includes(field.name.value))
-      .map(
-        (field): FieldDefinitionNode => ({
-          ...field,
-          description: undefined,
-          directives:
-            federationVersion === FederationVersion.V1
-              ? [externalDirective]
-              : [],
-        })
-      ) || []
-  );
+  const allfields = nodes.flatMap((node) => node.fields || []);
+  return (keyFieldNames
+    .map((keyFieldName) => {
+      const matchingField = allfields.find(
+        (field) => field.name.value === keyFieldName
+      );
+      return {
+        ...matchingField,
+        description: undefined,
+        directives:
+          federationVersion === FederationVersion.V1 && nodeIsEntity
+            ? [externalDirective]
+            : [],
+      };
+    })
+    .filter(Boolean) || []) as FieldDefinitionNode[];
 }
 
 /**
@@ -224,30 +244,46 @@ function generateComplexKeyObjectTypes(
     if (keyMapping[key]) {
       const currentField = fields.find((field) => field.name.value === key);
       const fieldType = extractFieldType(currentField);
-      const currentNode = definitionNodes.find(
-        (node) => node.name.value === fieldType
-      );
+      const currentNodes =
+        definitionNodes.filter((node) => node.name.value === fieldType) || [];
+      const currentNode = currentNodes[0];
 
-      if (!currentNode) {
+      if (!currentNodes.length) {
         return;
       }
 
+      const nodeIsEntity = isEntity(currentNodes);
+      const existingNode = objectTypes[fieldType];
+      const existingFields = existingNode?.fields || [];
+      const existingDirectives = existingNode?.directives
+        ? {directives: existingNode.directives}
+        : {};
+      const existingInterfaces = existingNode?.interfaces
+        ? {interfaces: existingNode.interfaces}
+        : {};
+
       const subKeyFields = getKeyFields(
-        currentNode,
+        currentNodes,
         keyMapping[key] || {},
         federationVersion
       );
 
-      if (!objectTypes.hasOwnProperty(fieldType)) {
-        objectTypes[fieldType] = {
-          kind:
-            federationVersion === FederationVersion.V1
-              ? Kind.OBJECT_TYPE_EXTENSION
-              : Kind.OBJECT_TYPE_DEFINITION,
-          name: currentNode.name,
-          fields: subKeyFields,
-        };
-      }
+      const deduplicatedSubKeyFields = subKeyFields.filter((field) =>
+        existingFields.every(
+          (existingField) => existingField.name.value !== field.name.value
+        )
+      );
+
+      objectTypes[fieldType] = {
+        kind:
+          federationVersion === FederationVersion.V1 && nodeIsEntity
+            ? Kind.OBJECT_TYPE_EXTENSION
+            : Kind.OBJECT_TYPE_DEFINITION,
+        name: currentNode.name,
+        fields: [...existingFields, ...deduplicatedSubKeyFields],
+        ...existingDirectives,
+        ...existingInterfaces,
+      };
 
       generateComplexKeyObjectTypes(
         definitionNodes,
@@ -267,12 +303,13 @@ export enum FederationVersion {
 
 export type GenerateRelayServiceSchemaOptions = {
   contractTags?: string[];
-  federationVersion?: FederationVersion;
+  federationVersion?: string;
   typeExceptions?: string[];
   nodeQualifier?: (
     node: DefinitionNode,
     objectTypes: Record<string, ObjectTypeNode>
   ) => boolean;
+  keySorter?: KeySorter;
 };
 
 /**
@@ -292,8 +329,23 @@ export function generateFroidSchema(
   options: GenerateRelayServiceSchemaOptions = {}
 ): DocumentNode {
   // defaults
-  const federationVersion = options?.federationVersion ?? FederationVersion.V2;
+  const explicitFederationVersion =
+    options?.federationVersion ?? FED2_DEFAULT_VERSION;
+
+  if (
+    explicitFederationVersion !== FED1_VERSION &&
+    explicitFederationVersion.indexOf(FED2_VERSION_PREFIX) === -1
+  ) {
+    throw new Error(
+      `Federation version must be either '${FED1_VERSION}' or a valid '${FED2_VERSION_PREFIX}x' version. Examples: v1, v2.0, v2.3`
+    );
+  }
+  const federationVersion =
+    explicitFederationVersion === FED1_VERSION
+      ? FederationVersion.V1
+      : FederationVersion.V2;
   const typeExceptions = options?.typeExceptions || [];
+  const keySorter = options?.keySorter || defaultKeySorter;
   const nodeQualifier = options?.nodeQualifier || (() => true);
   const allTagDirectives: ConstDirectiveNode[] =
     options?.contractTags?.sort().map((tag) => createTagDirective(tag)) || [];
@@ -334,7 +386,7 @@ export function generateFroidSchema(
           return objectTypes;
         }
 
-        const validKeyDirective = selectValidKeyDirective(node);
+        const validKeyDirective = selectValidKeyDirective(node, keySorter);
 
         if (!validKeyDirective) {
           return objectTypes;
@@ -343,7 +395,7 @@ export function generateFroidSchema(
         const {keyMappingRecord, keyDirective} = validKeyDirective;
 
         const keyFields = getKeyFields(
-          node,
+          [node],
           keyMappingRecord,
           federationVersion
         );
@@ -353,15 +405,30 @@ export function generateFroidSchema(
           extensionAndDefinitionNodes
         );
 
+        const nodeIsEntity = isEntity([node]);
+        const existingNode = objectTypes[node.name.value];
+        const existingFields = (existingNode?.fields || []).filter(
+          (field) => field.name.value !== ID_FIELD_NAME
+        );
+        const dedupedKeyFields = keyFields.filter((field) =>
+          existingFields.every(
+            (existingField) => existingField.name.value !== field.name.value
+          )
+        );
+
         objectTypes[node.name.value] = {
           kind:
-            federationVersion === FederationVersion.V1
+            federationVersion === FederationVersion.V1 && nodeIsEntity
               ? Kind.OBJECT_TYPE_EXTENSION
               : Kind.OBJECT_TYPE_DEFINITION,
           name: node.name,
-          interfaces: [implementsNodeInterface],
-          directives: [keyDirective],
-          fields: [createIdField(idTagDirectives), ...keyFields],
+          interfaces: existingNode?.interfaces || [implementsNodeInterface],
+          directives: existingNode?.directives || [keyDirective],
+          fields: [
+            createIdField(idTagDirectives),
+            ...existingFields,
+            ...dedupedKeyFields,
+          ],
         };
 
         generateComplexKeyObjectTypes(
@@ -380,7 +447,10 @@ export function generateFroidSchema(
   const tagDefinition =
     federationVersion === FederationVersion.V1
       ? createFederationV1TagDirectiveDefinition()
-      : createLinkSchemaExtension(['@key', '@tag']);
+      : createLinkSchemaExtension(
+          ['@key', '@tag'],
+          explicitFederationVersion as string
+        );
 
   // build schema
   return {